PROJ-189-MongoDB-schema-and-data-migration

+++ src/Web/services/labelService.js
@@ -1,4 +1,5 @@
 const Label = require('../models/labels');
+const { removeLabelFromMail } = require('./mailService');
 
 // Create a new label associated with a user
 async function createLabel(userObjectId, name, color) {
@@ -27,6 +28,19 @@ async function updateLabelById(labelId, userObjectId, updates) {
 
 // Delete label by _id and user
 async function deleteLabelById(labelId, userObjectId) {
+  const label = await Label.findOne({ _id: labelId, user: userObjectId });
+  if (!label) return null;
+
+  const mailIds = label.mailIds || [];
+
+  // Remove label from each mail using the existing logic
+  await Promise.all(
+    mailIds.map(mailId =>
+      removeLabelFromMail(userObjectId, mailId.toString(), labelId.toString())
+    )
+  );
+
+  // Finally delete the label itself
   return await Label.findOneAndDelete({ _id: labelId, user: userObjectId });
 }
